housekeeping and stylization changes

+++ src/price.cpp
@@ -4,11 +4,11 @@
 
 price::price() : price(-1, 0) {}
 
-price::price(int pri, uint32_t timestamp) {
+price::price(int pri, uint64_t timestamp) {
   update(pri, timestamp);
 }
 
-void price::update(int pri, uint32_t timestamp) {
+void price::update(int pri, uint64_t timestamp) {
   this->pri = pri;
   this->timestamp = timestamp;
 }
@@ -17,6 +17,6 @@ int price::get_price() {
   return this->pri;
 }
 
-uint32_t price::get_timestamp() {
+uint64_t price::get_timestamp() {
   return this->timestamp;
 }
